+++ fabric/src/main/java/io/github/sakurawald/fuji/module/initializer/command_event/CommandEventInitializer.java
@@ -58,6 +58,7 @@ public class CommandEventInitializer extends ModuleInitializer {
         var onPlayerFirstJoinedConfig = CommandEventInitializer.config.model().event.on_player_first_joined;
         if (onPlayerFirstJoinedConfig.enable) {
             // NOTE: If you use `Stats.LEAVE_GAME < 1` as the stat value, then it will not get saved when the server is stopped by `/stop`.
+            // The vanilla Minecraft thinks the `dis-connect` by the server is not identical to `leave the game` by the player.
             int stat = player.getStatHandler().getStat(Stats.CUSTOM.getOrCreateStat(Stats.TOTAL_WORLD_TIME));
             if (stat == 0) {
                 CommandEventInitializer.executeCommandOnEvent(player, onPlayerFirstJoinedConfig.command_list);