+++ CHANGELOG.md
@@ -1,3 +1,4 @@
 > Read detailed change logs in https://github.com/sakurawald/fuji/commits/dev/
 
 - [command_event] fix: the `on_player_first_join_event` will be trigger again for a player, if the server is shutdown by `/stop` before the `LEAVE_GAME` stat gets `updated` and `saved` (#432)
+- [command_bundle] feature: improve the `command feedback` for `/command-bundle {register/unregister}` command.